Add problem text

+++ Ordered list generator/Ordered list generator.cpp
@@ -1,3 +1,21 @@
+// 
+// Problem text: Using the Standard Library and C++(03, 11 or 14) write a function with the following signature that 
+// generates a list of random numbers between 1 and 1000, sort them comparing the remainder of a division by 11 and 
+// starting from the bigger remainder, and print them on the standard output.
+// 
+// For instance:
+// 
+//  void PrintRndByEleven(int n /* numbers to generate */)
+//  {
+// 	 // Your code goes here
+//  }
+// 
+//  PrintRndByEleven(5) could generate 741, 985, 168, 73, 12 and print “73, 985, 741, 168, 12”
+// 
+// 
+
+
+
 #include <iostream>
 #include <vector>
 #include <algorithm>
@@ -9,7 +27,7 @@ void PrintRndByEleven(const int n /* generate n numbers */)
 	static bool firstCall = true;
 	if (firstCall)
 	{
-		std::srand(time(0));	// Make numbers more random - call this once when the program starts
+		std::srand(static_cast<unsigned int>(time(0)));	// Make numbers more random - call this once when the program starts
 		firstCall = false;
 	}
